+++ src/Hst.Imager.Core/Commands/RdbPartExportCommand.cs
@@ -77,8 +77,8 @@
             // calculate source cylinder size, offset and size
             var sourceCylinderSize = sourceRigidDiskBlock.Heads * sourceRigidDiskBlock.Sectors *
                                      sourceRigidDiskBlock.BlockSize;
-            var sourceOffset = partitionBlock.LowCyl * sourceCylinderSize;
-            var sourceSize = (partitionBlock.HighCyl - partitionBlock.LowCyl + 1) * sourceCylinderSize;
+            var sourceOffset = (long)partitionBlock.LowCyl * sourceCylinderSize;
+            var sourceSize = ((long)partitionBlock.HighCyl - partitionBlock.LowCyl + 1) * sourceCylinderSize;
 
             OnDebugMessage($"Opening destination path '{destinationPath}' as writable");
